rename util to utils part 2
