+++ napari/_vispy/__init__.py
@@ -1,1 +1,1 @@
-from .util import create_vispy_visual
+from .utils import create_vispy_visual
