+++ napari/components/tests/test_multichannel.py
@@ -1,7 +1,7 @@
 import numpy as np
 import dask.array as da
 from napari.components import ViewerModel
-from napari.util import colormaps
+from napari.utils import colormaps
 
 base_colormaps = colormaps.CYMRGB
 two_colormaps = colormaps.MAGENTA_GREEN
