+++ napari/components/viewer_model.py
@@ -6,12 +6,12 @@ from xml.etree.ElementTree import Element, tostring
 from .dims import Dims
 from .layerlist import LayerList
 from .. import layers
-from ..util import colormaps
-from ..util.event import EmitterGroup, Event
-from ..util.keybindings import KeymapMixin
-from ..util.theme import palettes
-from ..util.misc import ensure_iterable, is_iterable
-from ..util import io
+from ..utils import colormaps
+from ..utils.event import EmitterGroup, Event
+from ..utils.keybindings import KeymapMixin
+from ..utils.theme import palettes
+from ..utils.misc import ensure_iterable, is_iterable
+from ..utils import io
 
 
 class ViewerModel(KeymapMixin):
