+++ napari/layers/base/_constants.py
@@ -1,6 +1,6 @@
 from enum import auto
 
-from ...util.misc import StringEnum
+from ...utils.misc import StringEnum
 
 
 class Blending(StringEnum):
