+++ napari/layers/image/image.py
@@ -9,10 +9,10 @@ from scipy import ndimage as ndi
 import vispy.color
 from ..base import Layer
 from ..layer_utils import calc_data_range, increment_unnamed_colormap
-from ...util.event import Event
-from ...util.status_messages import format_float
+from ...utils.event import Event
+from ...utils.status_messages import format_float
 from ._constants import Rendering, Interpolation
-from ...util.colormaps import make_colorbar, AVAILABLE_COLORMAPS
+from ...utils.colormaps import make_colorbar, AVAILABLE_COLORMAPS
 from .image_utils import get_pyramid_and_rgb
