+++ napari/layers/shapes/shape_models/shape.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 import numpy as np
 from copy import copy
 from vispy.color import Color
-from ..shape_util import (
+from ..shape_utils import (
     triangulate_edge,
     triangulate_face,
     is_collinear,
