+++ napari/layers/vectors/vectors.py
@@ -3,9 +3,9 @@ from xml.etree.ElementTree import Element
 import numpy as np
 from copy import copy
 from ..base import Layer
-from ...util.event import Event
-from ...util.status_messages import format_float
-from .vectors_util import vectors_to_coordinates, generate_vector_meshes
+from ...utils.event import Event
+from ...utils.status_messages import format_float
+from .vector_utils import vectors_to_coordinates, generate_vector_meshes
 from vispy.color import get_color_names, Color
